compat-wireless: copy driver files only if present
This modifies admin-update.sh to only copy files over
from a driver target directory if files are found. This
should avoid a bunch of complaints of copying files
from directories that are just placeholders for other
folders, as happens with drivers/net/wireless/brcm80211/
which has folders a Makefile, Kconfig but no files there
itself:

	brcmfmac/
	brcmsmac/
	brcmutil/
	include
	Kconfig
	Makefile

+++ scripts/admin-update.sh
@@ -350,9 +350,18 @@ done
 # Drivers in their own directory
 for i in $DRIVERS; do
 	mkdir -p $i
-	echo "Copying $GIT_TREE/$i/*.[ch]"
-	cp $GIT_TREE/$i/*.[ch] $i/
-	cp $GIT_TREE/$i/Makefile $i/
+
+	# -print -quit will just print once, so we don't burden
+	# this script with searching for further files if one is
+	# found
+	FILES_FOUND=$(find $GIT_TREE/$i/ -maxdepth 1 -type f -name \*.[ch] -print -quit | wc -l)
+	if [ $FILES_FOUND -eq 1 ]; then
+		echo "Copying $GIT_TREE/$i/*.[ch]"
+		cp $GIT_TREE/$i/*.[ch] $i/
+	fi
+	if [ -f $GIT_TREE/$i/Makefile ]; then
+		cp $GIT_TREE/$i/Makefile $i/
+	fi
 	if [ -f $GIT_TREE/$i/Kconfig ]; then
 		cp $GIT_TREE/$i/Kconfig $i/
 	fi
